+++ blank_line_after_blocks/__init__.py
@@ -1,3 +1,5 @@
 """A Python formatter to add blank lines after code blocks."""
 
-__version__ = '0.1.0'
+import importlib.metadata
+
+__version__ = importlib.metadata.version('blank-line-after-blocks')