+++ blank_line_after_blocks/main_py.py
@@ -7,6 +7,7 @@ from pathlib import Path
 import click
 
 import blank_line_after_blocks.helper as helper
+from blank_line_after_blocks import __version__
 from blank_line_after_blocks.base_fixer import BaseFixer
 
 
@@ -54,6 +55,7 @@ class PythonFileFixer(BaseFixer):
 
 
 @click.command()
+@click.version_option(version=__version__)
 @click.argument('paths', nargs=-1, type=click.Path())
 @click.option(
     '--exclude',